+++ src/combat.h
@@ -2,9 +2,9 @@
 #include "loot_table.h"
 
 struct Combat {
-    float       hitPointsMax    {};
-    float       hitPoints       {};
-    float       meleeDamage     {};  // TODO: add min/max and randomly choose?
+    int         hitPointsMax    {};
+    int         hitPoints       {};
+    int         meleeDamage     {};  // TODO: add min/max and randomly choose?
     double      attackStartedAt {};
     double      attackDuration  {};
     uint32_t    attackFrame     {};
