+++ src/healthbar.h
@@ -5,7 +5,7 @@
 
 struct HealthBar {
     static void SetFont(const Font font);
-    static void Draw(int fontSize, const Sprite &sprite, const Body3D &body, const char *name, float hitPoints, float maxHitPoints);
+    static void Draw(int fontSize, const Sprite &sprite, const Body3D &body, const char *name, int hitPoints, int maxHitPoints);
 
 private:
     static Font s_font;
